Normal buffer is 8-bit.

+++ samples/_opengl/DeferredShadingAdvanced/src/DeferredShadingAdvancedApp.cpp
@@ -679,7 +679,7 @@ void DeferredShadingAdvancedApp::resize()
 														 .dataType( GL_BYTE ) );
 		mTextureFboGBuffer[ 2 ]	= gl::Texture2d::create( windowSize.x, windowSize.y, 
 														 gl::Texture2d::Format()
-														 .internalFormat( GL_RG16F )
+														 .internalFormat( GL_RG8 )
 														 .magFilter( GL_NEAREST )
 														 .minFilter( GL_NEAREST )
 														 .wrap( GL_CLAMP_TO_EDGE )
